Use variables in config, remove versions in file names
There was a lot of repeated information in the config for the virtualbox and vmware configuration. Moving it to variables reduces it down to a single placee to update.

There will only be one version of ubuntu or debian in the repository, looking to the iso_url variable in the config will identify the version. this will reduce renaming requirements. If you want to build an old version check out an older commit.