Short script names for re-use. Shorten platform names for re-use.
There will only ever be 64bit ubuntu and debian distributions for springboard. Specific versions will be the latest stable at the time of the springboard build. It should not be necessary to reference the distribution versions in the files. It will also save re-naming ad infinum as new versions are released.